DP-5382 - Replace complex join with two separate queries.
The complex join returns in some cases no results. This happens if there are no lines
in planes with kosten_base = 'baten' buth there are lines with 'kosten'.

To fix this it is more clear to split this up in two separate queries. The first checks if
the plannr exists and gets general properties. If this does not exist return 404

The second gets the data from planex and aggregates the 'baten' If this does not return
results  only this part of the returned data is set to None